fix(rising-seasons): rename app, add filters, fix colors, fix The Office
Renames apps/imdb-rising → apps/rising-seasons throughout (npm scripts,
GH workflow, internal docs, canonical URL, storage namespace). Public
title and tagline drop "IMDb" to stay generic about the data source.

The Office fix:
- Findmatches now does two passes: first identifies series with at least
  one shape-matching season, then emits ALL seasons of those series.
  Seasons that don't fit a curve pattern get shapes: [] but stay
  searchable. The Office (US) now shows all 9 seasons (1, 6, 7, 8 carry
  a "No pattern" tag) instead of the 5 that match a shape outright.
- data.json grew 6.0 MB → 10.7 MB; 4,230 → 7,598 seasons.
- Updated tests to cover the new behaviour (shape-less seasons of
  qualifying series are now included).

5 new filters (was 7 input filters, now 12):
- Max episodes — companion to min episodes for short seasons
- Min average rating — skip mediocre seasons
- Min climb — only seasons that rose at least N points
- Series type — TV series / mini-series / all
- Poster — any / with poster / without poster

UI polish:
- Subtitle replaced: "Built X · Y seasons indexed across all of IMDb"
  → "Last updated: May 8, 2026"  (formatted from data.builtAt)
- Card stats relabeled: "Avg 7.5" (was "avg 7.5") and
  "1,000 votes/ep min" (was "1,000+ votes/ep" — the trailing + read as
  a stray crossbar next to the average rating)
- Modal episode list shows ratings as plain numbers (no leading star)

Dark dropdowns:
- color-scheme: dark on :root and on every <select> so the native popup
  menus render dark, not the OS-default light. <option> elements
  explicitly themed for browsers (Firefox) that ignore color-scheme on
  options.

Button text colours:
- main.css line 322 has `button:hover { color: #ce1b28 !important }`
  which was turning every button text red on hover. Every button-class
  rule (.btn / .btn-ghost / .shape-chip / .genre-chip / .view-btn /
  .modal-close / .watch-btn / .watch-toggle / .page-btn) now sets
  `color: ... !important` for both default and :hover so the
  class-scoped colour wins.
- Genre chips and "Reset all filters" now use white text by default
  (were muted gray).
- Disabled pagination buttons keep white text at lower opacity instead
  of fading to a separate muted shade.

apps.html:
- New section listing Rising Seasons with the same content/icon/button
  pattern as Mario Kart, Gym Tracker, and Football H2H.
