+++ apps.html
@@ -165,6 +165,18 @@
             </ul>
           </div>
         </section>
+        <section>
+          <div class="content">
+            <header>
+              <a href="apps/rising-seasons/index.html" class="icon fa-line-chart brand-color-primary" aria-label="Open Rising Seasons"></a>
+              <h3>Rising Seasons</h3>
+            </header>
+            <p>Discover TV shows by the shape of their episode ratings. Browse rising, consistent, slow-burn, big-finale, or rebound seasons across all of TV.</p>
+            <ul class="actions special">
+              <li><a href="apps/rising-seasons/index.html" class="button">Browse Seasons</a></li>
+            </ul>
+          </div>
+        </section>
       </div>
     </div>
   </main>
