--- apps/imdb-rising/.gitignore
@@ -1,1 +1,0 @@
-# data.json is now committed and refreshed weekly by .github/workflows/refresh-imdb.yml
