--- apps/imdb-rising/data/.gitignore
@@ -1,3 +1,0 @@
-*.tsv
-*.tsv.gz
-tmdb-cache.json
