+++ apps/rising-seasons/.gitignore
@@ -1,0 +1,1 @@
+# data.json is now committed and refreshed weekly by .github/workflows/refresh-imdb.yml
