+++ apps/rising-seasons/data/.gitignore
@@ -1,0 +1,3 @@
+*.tsv
+*.tsv.gz
+tmdb-cache.json
