+++ package.json
@@ -3,11 +3,11 @@
   "private": true,
   "type": "commonjs",
   "scripts": {
-    "test": "node --test apps/gym-tracker/tests/ apps/football-h2h/tests/ apps/imdb-rising/tests/",
+    "test": "node --test apps/gym-tracker/tests/ apps/football-h2h/tests/ apps/rising-seasons/tests/",
     "test:gym": "node --test apps/gym-tracker/tests/",
     "test:football": "node --test apps/football-h2h/tests/",
-    "test:imdb-rising": "node --test apps/imdb-rising/tests/",
-    "build:imdb-rising": "node apps/imdb-rising/scripts/build-data.js",
-    "enrich:imdb-rising": "node apps/imdb-rising/scripts/enrich-tmdb.js"
+    "test:rising-seasons": "node --test apps/rising-seasons/tests/",
+    "build:rising-seasons": "node apps/rising-seasons/scripts/build-data.js",
+    "enrich:rising-seasons": "node apps/rising-seasons/scripts/enrich-tmdb.js"
   }
 }
